add chocolate and peanut topping

+++ src/main/java/Cake.java
@@ -1,4 +1,6 @@
+import java.math.BigDecimal;
+
 public interface Cake {
     String getName();
-    Long getPrice();
+    BigDecimal getPrice();
 }
